Verify Return level accuracy + Refactor Return level class
I have made several modifications to the Return level class so that it is found directly from the fit class and not in the base GEV classes. This allows for better usability. I also checked for Return levels behaviour and fit behaviour. Fit behaviour is confirmed mathamtically but return level not quite yet. Still has some investigations to do.